+++ pkg/apihandler/apitransport.go
@@ -241,7 +241,8 @@ func (t *ApiTransport) internalGraphQLRoundTrip(request *http.Request) (res *htt
 		}
 	}
 
-	if clientRequest := request.Context().Value(pool.ClientRequestKey).(*http.Request); clientRequest != nil {
+	// Make sure we account for both pool.ClientRequestKey being nil and being non present
+	if clientRequest, ok := request.Context().Value(pool.ClientRequestKey).(*http.Request); ok && clientRequest != nil {
 		requestJSON, err := hooks.HttpRequestToWunderGraphRequestJSON(clientRequest, false)
 		if err != nil {
 			return nil, err