+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.4...@wundergraph/swr@0.13.5) (2023-03-06)
+
+### Bug Fixes
+
+* upload types ([#723](https://github.com/wundergraph/wundergraph/issues/723)) ([5077a7d](https://github.com/wundergraph/wundergraph/commit/5077a7d2ab363d3bbbb1df35423e5ad799b1800f)) (@Pagebakers)
+
 ## [0.13.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.3...@wundergraph/swr@0.13.4) (2023-03-01)
 
 **Note:** Version bump only for package @wundergraph/swr